Initial import of GDBus++ (gdbuspp)
The GDBus++ code base has roots in the OpenVPN 3 Linux project, where
the aim is to provide a much better implementation of the C++ interface
wrapping of the glib2 GDBus functionality.

This project has quite some code migrated from the OpenVPN 3 Linux
project, as well as new code to make this a better C++ implementation
from the starting point.

A compiler with C++17 is a requirement to build this project.

Please read the README.md file for more details.

Signed-off-by: David Sommerseth <davids@openvpn.net>