service: Make Service::Create() return std::shared_ptr<C>
The DBus::Service::Create() static function was set up to return
DBus::Service::Ptr.  This makes it harder to extend the Service class in
the implementation and make use of that functionality.

By return a ::Ptr (aka std::shared_ptr) of the class being
instantiated instead of DBus::Serivce, all the functionality is
available directly.

Signed-off-by: David Sommerseth <davids@openvpn.net>