credentials: New component, for querying credential details
The DBus::Credentials namespace is added with a Query component.  This
is used to retrieve information about a specific D-Bus bus name, like
the process ID (pid) or the user ID (uid) which this bus name is running
under.  D-Bus services can also be queried for the unique bus name it
uses via the well-known bus name.

Signed-off-by: David Sommerseth <davids@openvpn.net>