signals/group: Rework the Signals::Group API slightly
Make the Signals::Group have a fixed object path and interface.  That is
used as information identifying the sender of the signal.

Then reduce the Signals::Group::AddTarget() to only take the
destination bus name where to send the signals.  This can be called
multiple times to send to more targets at the same time.

With this change, the D-Bus introspection Signals::Group generates will
also make a lot more sense, as this object is intended to be used
together with a DBus::Object::Base object.

Signed-off-by: David Sommerseth <davids@openvpn.net>