features: Implement basics for IdleDetect support
The IdleDetect features is not exposed directly to GDBus++
implementations, but used internally.  This feature is driven by the
DBus::Object::Manager class, which provides the public APIs to enable
this feature, together with a couple of support methods in
DBus::Object::Base.

When the IdleDetect feature is enabled in a D-Bus service, it will
shutdown the service if the service has been detected running idle for
specified time.  The idle time is based on the time since the last
incoming D-Bus call or if the DBus::Object::Manager has some
DBus::Object::Base objects registered.

Signed-off-by: David Sommerseth <davids@openvpn.net>