signals: Extend Signals::Group to support multiple distribution lists
The current implementation only had a single set of distribution targets
for the signals, configured using the AddTarget() method.  For
applications needing to send signals to different set of recipients in
various parts of the code, this implementation is too limited.

This is improved by adding distribution list groups, which extends the
API with methods to add/remove groups and to add recipients (aka
targets) to these groups.  Sending signals to these groups is handled
via the GroupSendGVariant() method.

To avoid reimplementing too much of existing logic, the "non-group"
interface is preserved but under the hood they use the "__default__"
group name, which is considered a reserved group name not to be used
outside this class.

Signed-off-by: David Sommerseth <davids@openvpn.net>