authz: The authorization request details should not be private
There is no way to easily access this information when performing an
authorization.  This information is const and is expected to be shared
with a DBus::Object::Base object, this does not need any other API
around it.

This also removes the need for the friend declaration to the Exception
class.

Signed-off-by: David Sommerseth <davids@openvpn.net>