object/property: Enhance the DBus::Object::Base::AddPropertyBySpec() API
This setup in a way where it required the implementer to always provide
both a get and set property callback function, even if the property is
read-only.

Instead, remove the readwrite argument and make two methods of the
AddPropertyBySpec() method. The read-write property type will have both
the get and set property callbacks, while the read-only type will only
have a single get property callback.  The boolean readwrite flag is
implicitly understood via the callbacks provided.

This logic is at this time only implemented in the top-level
DBus::Object::Base level.

Signed-off-by: David Sommerseth <davids@openvpn.net>