glib2/utils: Split up glib2::Value::Create()
There were two types of the Create() method; one which takes only a
single argument of the value to create the GVariant object, and the
other one which could take a data type argument as well.

This could be a bit confusing, so split this up into
glib2::Value::Create() which only takes a single argument and
glib2::Value::CreateType() which takes the data type in addition.

Signed-off-by: David Sommerseth <davids@openvpn.net>