object/property: Properties are not allowed to return nullptr
If a property value is nullptr, an error must be set [1].  This is
handled by the GetProperty() implementation throwing
DBus::Object::Property::Exception or DBus::Object::Exception.  Those
exceptions are handled correctly and will set the glib2 GError
correctly.

In all other cases, if a GetProperty() callback returns nullptr; handle
this more gracefully to avoid the glib2 library to shutdown the D-Bus
connection.

Signed-off-by: David Sommerseth <davids@openvpn.net>