signals: Signals::Emit() should not throw an exception on error
Instead of throwing an exception when it cannot send a signal, send the
information straight to stderr.  This is to avoid programs to
unexpectedly stop and enforcing the API to expect exception catching
will be to intrusive.  This is not necessarily a fatal error, even
though signals cannot be emitted.

To provide a way for the caller to know if it worked, change the API to
return a bool value indicating if the signals to all targets was emitted
successfully.

Signed-off-by: David Sommerseth <davids@openvpn.net>