glib2/callbacks: Allow DBus::Method::Exception to not trigger a program error
The current implementation will print an error message to stdout
whenever a DBus::Exception is thrown.  This has values since these
may indicate a program/implementation error.

But an implementation may also throw and exception to be targeting
the calling user; these are not intended to be program errors but
"user errors".

This change will send DBus::Method::Exception only to the end-user
(via the D-Bus error reporting mechanisms) but will not log this as
an error for the program.

Signed-off-by: David Sommerseth <davids@openvpn.net>