tests: Add regression test for glib2::Utils::checkParams()
A bug has been found in glib2::Utils::checkParams() causing misbehaviour
when the input parameters are not of a container type bu the number of
elements value is larger than 0.

This test is added here to verify it breaks with the current code and
the coming commit fixing it fixes this issue.

Signed-off-by: David Sommerseth <davids@openvpn.net>