glib2/utils: Fix checkParams() validation on non-container types
This function calls g_variant_n_children() to check how many child
elements the input parameters to check has.  On non-container types,
there are no children and the g_variant_n_children() call will fail and
may even crash the running program.

Add a check that the input parameters are of a container type before
extracting number of children.  In addition require that the "num"
argument of the function must check a container type if "num" is higher
than 0.

Signed-off-by: David Sommerseth <davids@openvpn.net>