AsyncProcess::Pool: Fix unexpected behaviour on single core systems
When creating the GThreadPool, the limitation of number of worker
threads are set to half of available CPU cores.  On a single core
system, this ends up to be 0.

The g_thread_pool_new() will in this case actually create a
GThreadPool and g_thread_pool_push() will happily accept new tasks.
But since the number of worker threads are 0, none of these tasks will
ever be executed.

Ensure that there is at least 1 worker thread available.

Signed-off-by: David Sommerseth <davids@openvpn.net>