tests: Make test-data-types independent of simple-service
The test-data-types test program needed the simple-service to be
available for one of its tests.  For build environments, the D-Bus
daemon might not be available so the simple-service program will not
start.

Split this into two run modes, with or without testing simple-service
being available.

This allows some basic tests to be run in the automated build
environments even if the D-Bus framework is otherwise not available.

Signed-off-by: David Sommerseth <davids@openvpn.net>