build/tests: Rework the test framework setup
This modifies the whole setup to explicitly enable the simple-service
kick-off process only on the test cases which need it instead of using
the add_test_setup() which would start the simple-service regardless if
the test case needed it or not.

With this change, a few of the tests can now be run automatically in a
'standalone' mode when run within automated build environments where the
D-Bus framework is not running and available.  In addition, these
standalone tests can also easily run in parallel.

To run only the standalone tests:

    $ meson test --suite standalone

Signed-off-by: David Sommerseth <davids@openvpn.net>