signals: Add Signals::Signal API
This is a slightly different approach where a single D-Bus signal is
implemented as a single C++ object, inheriting the Signals::Signal
class.

This allows this object to declare the signal arguments (D-Bus data
type) internally, which will be validated before sending the signal.

See the tests/signal-signal.cpp test program for details.

Signed-off-by: David Sommerseth <davids@openvpn.net>