signals/group: Extend Signal::Group with CreateSignal<>()
The new CreateSignal<>() and GroupCreateSignal<>() methods in
Signals::Group will create a new Signals::Signal based object and
configure it with the appropriate Signals::Emit object to send the
signal with the right targets.

The tests/signal-group test program has reimplemented the "Debug" signal
to use this approach for testing.

Signed-off-by: David Sommerseth <davids@openvpn.net>