signals/manager: Clean up properly when SubscriptionManager is removed
When the SubscriptionManager is deleted, the D-Bus signal must also be
unsubscribed via the g_dbus_connection_signal_unsubscribe() function.

If this is not done and the program still runs, glib2 will continue to
attempt to call the callback method assigned for the subscription.  But
since the SubscriptionManager object and all the SingleSubscription
objects it owns are deleted, the callback function is no long existing.
This results in a segfault situation in the program.

Signed-off-by: David Sommerseth <davids@openvpn.net>