proxy: Further enhance DBus::Proxy::Client setup checks
Commit 1e4bebabb9bec963e582a646c2c9e8797c446ed6 tried to improve the
DBus::Proxy::Client object setup phase to ensure the service was
reachable.  This turned out to cause several failures and was also less
effective.

This new approach will instead query the main D-Bus daemon the
DBus::Connection is connected to, to check if there is a service name
available with a given name and if not, check if it can be started
before trying to starting it.

With this change, we are also in charge of the error message returned
back to the caller if it fails.

Signed-off-by: David Sommerseth <davids@openvpn.net>