service: Remove the obscure GDBusConnection * callback argument
When implementing up DBus::Service, the BusNameAcquired() and
BusNameLost() callback methods needed to receive a GDBusConnection *
argument.

This argument should not be needed; this is already available via the
Connection::Ptr passed to the DBus::Service constructor.  This
connection should not be any different from what is being passed via
callback.  If the raw GDBusConnection pointer would ever be needed, that
is available via the Connection::ConnPtr() method

glib2 passes this pointer due to the C approach is more restricted in
storing such information and making it available for the callback
function itself.  But in C++, the callback happens inside the
DBus::Service object - so we have all the information needed in the
object.

Passing this information is thus not needed in GDBus++ and was an API
oversight in during the early implementation phase.

Signed-off-by: David Sommerseth <davids@openvpn.net>