connection: Remove the Connection(GDBusConnection *) constructor
The Connection(GDBusConnection *) constructor is a bit odd, as it would
require you to have set up a D-Bus connection outside of GDBus++.  This
should not be needed and provides no real benefits.  In addition, it
will not provide details about the bus type (glib2 APIs does not provide
such a query mechanism).

Signed-off-by: David Sommerseth <davids@openvpn.net>