proxy/utils: Improve DBusServiceQuery::CheckServiceAvail() checks
This method would misbehave if the D-Bus service could not be
automatically started.  To avoid that, do a lookup of the requested
service can be automatically started or not before continuing - and take
that result into consideration in the next steps.

Also add a bit more comments to this logic, to better explain why these
various checks are done.

Signed-off-by: David Sommerseth <davids@openvpn.net>