object/manager: Extend with Object::Manager::AttachRemoveCallback()
This extends the DBus::Object::Manager with a new AttachRemoveCallback()
method.  The purpose here is to allow an implementer to trigger a
callback lambda function when a specific object is being removed.

In some cases D-Bus objects will have the possibilities to "remove
themselves", but a parent object which created these objects will not
know such an event happened.  This allows these parent objects to get a
notification when this child object is being removed and can act upon
that event.

Signed-off-by: David Sommerseth <davids@openvpn.net>