exceptions: Make the exception error messages more end-user friendly
This tries to identify proxy errors and extracts the most useful parts
for the end-user.  The technical details will be available via the
error_domain in these exceptions.

This change is expected to be more visible when hit via the
DBus::Proxy::Exception paths - but implemented on the most central place
for now.

Signed-off-by: David Sommerseth <davids@openvpn.net>