proxy: Implement result callback for do_call_no_response()
The internal glib2::Proxy::do_call_no_response() method didn't set any
callback function when the asynchronous call was ready.  This is fine by
the glib2/gio documentation.  However, if the method being called
provides a result in return, it triggers an error from the dbus-daemon:

     [system] Rejected send message, 0 matched rules; type="method_return", sender=":1.1355" [...]

To avoid such noise in the system logs, a dummy callback which just
receives and frees the result is added.

Signed-off-by: David Sommerseth <davids@openvpn.net>