proxy/utils: CheckServiceAvail() must always call GetNameOwner()
The DBus::Proxy::Utils::DBusServiceQuery::CheckServiceAvail() method
would do a quick exit if the StartByService() call returned 2 - which
indicates the service is already running.

While the service may be running, it may not yet be fully settled.  It
merely means there is a registered bus name.  Do the GetNameOwner() as
well, ensure there is a relation between the well-known and unique bus
names.

Signed-off-by: David Sommerseth <davids@openvpn.net>