Fix -flto=auto link-time warning
Due to the #ifdef __G_IO_H__ check in gdbuspp/exceptions.hpp
we end up with conflicting types for DBus::Exception: one, with
more entries in its vtable because the check succeeds (in the
compiled library), and one with less because the check fails
(in the tests under docs/).

The problem can be seen when we setup meson to use -flto=auto:

meson setup -Dcpp_args="-flto=auto" _builddir

This patch #includes gio/gio.h directly in gdbuspp/exceptions.hpp
and removes the #ifdef check, so everybody ends up with the same
vtable for DBus::Exception objects.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>