proxy/utils: Allow Proxy::Utils::ServiceVersion() some time to settle
This method would just bail out early if it couldn't reach the target
object.  Normally this isn't a problem, when the service is already
running.  But if the service needs to start, it might need some time to
have settled enough to serve requests.

This grants totally about 750ms for the service to settle, in 5 x 150ms
steps.

Signed-off-by: David Sommerseth <davids@openvpn.net>