glib2: Add glib2::DataType::DBus<std::byte>
This adds support for the D-Bus BYTE ('y') data type.

The use of std::byte for this type was chosen over uint8_t, since
uint8_t is often a typedef of unsigned char.  That makes some of the C++
behaviour a bit more unexpected, especially when used in streams.

std::byte has a more consistent and explicit behavoir and represent the
D-Bus BYTE type more clearly.

Signed-off-by: David Sommerseth <davids@openvpn.net>