proxy: Add more method/property name details to exceptions
When the internal glib2::Proxy() class was instantiated, it
could be hard to grasp the context it was being called in
outside of the GDBus++ code.

Add some "error_details" containing the D-Bus method name or
Get/Set operation on the property.  This gives more
information to the caller of the public DBus::Proxy::Client
methods.

Signed-off-by: David Sommerseth <davids@openvpn.net>