glib2/utils: Improve vector allocation in glib2::Values::ExtractVector()
Since the size of the new vector can be extracted from the GVariant
object, use that value to reserve an initial size of the std::vector
being returned.  This reduces the number of memory allocation operations
needed.

In addition, extend the test-data-types unit test to also more
extensively test the glib2::Values::ExtractVector() functionality with
different data types.

Signed-off-by: David Sommerseth <davids@openvpn.net>