proxy: Use CheckServiceAvail() in proxy setups
When preparing a new proxy and connecting it to a service, use the
CheckServiceAvail() method instead of LookupService(),
LookupActivatable() and StartServiceByName().

The CheckServiceAvail() method will more persistently try to ensure it
has a connection running.  The Proxy::Client object could otherwise
often fail if the service is not starting quickly enough.

Since CheckServiceAvail() will respond quickly if the service is already
running, it just adds the needed delay to ensure it can reach the
service when returning.

Signed-off-by: David Sommerseth <davids@openvpn.net>