proxy: Make the Proxy::Client::Call() methods handle parameters not being a tuple
All the arguments passed to a D-Bus method call need to be wrapped into
a tuple set.  This can be done automatically now inside the
Proxy::Client implementation, by using the new glib2::Value::TupleWrap()
function.

While this does kind of break the D-Bus semantics, it makes the overall
code more readable and easier to understand.  And argument is an
argument to method call.  Needing to explicitly remember to wrap it into
a tuple just makes the code needlessly harder to write.

Signed-off-by: David Sommerseth <dazo@eurephia.org>