object/method: Make Object::Method::Arguments::SetMethodReturn() tuple wrap data
The API for Object::Method::Arguments::SetMethodReturn() required the
returned data to always be tuple wrapped.  This was not necessarily
clear from the code.  This is also not always convenient nor intuitive.

This change will make use of the glib2::Value::TupleWrap() function to
wrap the data as needed automatically.  Which simplifies the code
implementing a D-Bus service.

The tests/simple-service.cpp implementation was also refactored a bit to
make less use of glib2 functions directly and rather use the wrappers
provided in glib2::Value and glib2::Builder.

Signed-off-by: David Sommerseth <dazo@eurephia.org>