glib2: Relocate Value::Dict to Dict and expand it
While the current approach of preparing a GVariant based dictionary
works, it's not necessarily the best approach.  The upstream glib2
library has dedicated functions for this, and we can wrap them more
nicely in GDBus++

While these dictionaries itself are GVariant based values, it seems
better to have it in its own main namespace and not located under the
glib2::Value namespace.  This makes the API clearer on simpler value
types than what dictionaries are.

Signed-off-by: David Sommerseth <dazo@eurephia.org>