glib2::Builder/glib2::Value usage refactoring
The glib2 utilities API can now basically cover basically everything in
the higher API levels this library provides.  This reduces the use of
the lower level native glib2 g_variant_*() type of functions to a bare
minimum and makes use the functions in the glib2::Builder and
glib2::Value namespaces instead.

Signed-off-by: David Sommerseth <dazo@eurephia.org>