glib2: Add glib2::Builder::AddKeyValue() functions
This is an alternative to glib2::Dict::Add().  Using glib2::Dict will
enforce the data types for the key and values to be 'a{sv}' - which is
as expected for an ordinary dictionary type; the key is a string and the
value can be of any data types.

For use cases where the implementer wants more control of the data types
being used, the glib2::Dict (including the low-level g_variant_dict*()
functions) APIs lacks the capability to adjust the data type of the
dictionary container.

By using the glib2::Builder approach instead, the data type can be
enforced for both the key and value elements.  Since this will require a
slightly different approach to add new elements to the GVariantBuilder
container object, glib2::Builder::AddKeyValue() was added.

Signed-off-by: David Sommerseth <dazo@eurephia.org>